==> notebooks/1.3 edgelist_to_adjlist.ipynb <==
# Generate adjacenccy list file from edgelist where:
# 1) each row index is the node and
# 2) each value on the row is the neighbors to that node
import pandas as pd
import dask.dataframe as dd
import time
import csv
from fun.fun import *

# Variables
edges_fn = "../data/edges.csv"
edges_total = 684_732_453 # hardcoded

# Read edges to dataframe
perc = 100
nrows=int(edges_total*perc/100)
print("reading edges ... ", end='')
start = time.time()
df = pd.read_csv(edges_fn, nrows=nrows)
end = time.time()
print("read {:_} lines (took {:.1f}s)".format(len(df), (end-start)))

# get set of all unique nodes
nodes = sorted(set(df['source']).union(set(df['target'])))
print("Found {:_} unique nodes".format(len(nodes)))

# Open csv file
file = open('../data/nodelist.csv', 'w', newline='')
writer = csv.writer(file)

# Generate edges for each
nodes_handled, total = 0, len(nodes)
for node in nodes:
    edges = df[(df['source'] == node) | (df['target'] == node)]
    neighbors = set(edges['source']).union(set(edges['target']))
    neighbors.remove(node)
    csv_row = [node] + list(neighbors)
    writer.writerow(csv_row)
    nodes_handled, perc = track_progress(total, nodes_handled, text='progress:', inc=10)
    """ if perc > 10:
        break """
file.close()
