==> notebooks/1.4 Analysing Hashtags.ipynb <==
# Functions:
# 1) Reads degree centrality values form csv, 
# 2) displays histograms and 
# 3) analyses power law fit
import numpy as np
import pandas as pd
import dask.dataframe as dd
import matplotlib.pyplot as plt
pd.set_option('display.width', 500)
import time
import statsmodels.api as sm
from fun.fun import *

# Variables
edges_fn = '../data/edges.csv'
dataset_fn = "../dataset/TweetsCOV19.tsv"
edges_prq = "../data/edges.parquet"

# -> IN : Load tweets dataframe
tw = get_filtered_tweets_dataframe(dataset_fn)
print("Loaded {:_} tweets".format(len(tw)))
tw

# Get number of hashtags for each tweet
hashtags = {}
total, i = len(tw), 0
for _, row in tw.iterrows():
    for ht in row['Hashtags']:
        hashtags[ht] = hashtags.get(ht, 0) + 1
    i, perc = track_progress(total, i)
    #if perc >= 5: break
print("\nDone")

# Sort keys by frequency and filter out hashtags with 1 associated tweet
keys_sorted = sorted(hashtags.keys(), reverse=True, key=lambda x: hashtags[x])
print("Hashtags used: {:_}".format(len(keys_sorted)))
keys_sorted = [x for x in keys_sorted if hashtags[x] > 1]
print("Hashtags with more than 1 tweet: {:_}".format(len(keys_sorted)))

# Create dataframe with hashtags, frequencies and edges
from math import comb
htt = pd.DataFrame()
htt['hashtag'] = keys_sorted
htt['freq'] = htt['hashtag'].apply( lambda ht: hashtags[ht] )
htt['edges'] = htt['freq'].apply( lambda x: comb(x, 2) )
print(len(htt))
htt.head(10)
