==> notebooks/2.0 Calculating Degree.ipynb <==
# Calculated graph degree centrality from edge list and saves result to csv.
import numpy as np
import pandas as pd
import dask.dataframe as dd
pd.set_option('display.width', 500)
import time
from fun.fun import *

# -> IN : PARQUET Read edges from parquet to dataframe
edges_prq = "../data/edges.parquet"
print("reading edges ... ", end='')
start = time.time()
df = dd.read_parquet(edges_prq)
end = time.time()
print("read {:_} lines (took {:.1f}s)".format(len(df), (end-start)))
df.head()

# DATAFRAME : Calculate degree centrality values
source_counts = df['source'].value_counts().compute()
target_counts = df['target'].value_counts().compute()
sum_counts = source_counts.add(target_counts, fill_value=0).astype('int64')
print("Calculated degree centrality for {:_} nodes:".format(len(sum_counts)))
dc_df = pd.DataFrame({'node' : sum_counts.index, 'degree' : sum_counts.values})
dc_df.head()

# <- OUT : Save degree centrality values
dc_fn = '../data/degree_centralities.csv'
dc_df.to_csv(dc_fn, index=False)
