==> notebooks/2.test Dask and iGraph Testing.ipynb <==
import numpy as np
import pandas as pd
import scipy as sp
import dask.dataframe as dd
import dask.array as da
import igraph as ig
import matplotlib.pyplot as plt
import time
from fun.fun import *

# START
edges_csv = "../data/edges.csv"
edges_prq = "../data/edges.parquet"
edges_total = 684_732_453 # hardcoded

# -> IN : PARQUET Read edges from parquet to dataframe
print("reading edges ... ", end='')
start = time.time()
df = dd.read_parquet(edges_prq)
df[['source', 'target']] = df[['source', 'target']].astype('int32')
end = time.time()
print("read {:_} lines (took {:.1f}s)".format(len(df), (end-start)))
print(df.dtypes)
df.head()

# -> IN : Read list of nodes
nodes = pd.read_csv('../data/node_ids.csv', index_col='index')
print("Number of nodes: {:_}".format(len(nodes)))

# (1) GRAPH (ALL EDGES) : Loading edges into iGraph
print("Creating iGraph ...")
g = ig.Graph()
g.add_vertices(len(nodes))
print("Processing {} partitions ...".format(df.npartitions))
for i, partition in enumerate(df.to_delayed()):
    p_df = partition.compute()
    g.add_edges(p_df.values.tolist())
    _, perc = track_progress(df.npartitions, i, text="partitions processed:")
    #if perc > 1: break
    #if input("...") == 'b': break
    #if i+1 >= 10: break
print("\nDone.")
print("\nGRAPH CREATED:")
print("  nodes: {:_}".format(len(g.vs)))
print("  edges: {:_}".format(len(g.es)))

# (2) GRAPH (N EDGES) : Loading edges into iGraph
nrows = 4_000_000 # max: 4557647
#nrows = 100 # max: 4557647
g = ig.Graph.TupleList(df.head(nrows).values.tolist())
print("\nGRAPH CREATED:")
print("  nodes: {:_}".format(len(g.vs)))
print("  edges: {:_}".format(len(g.es)))

# Find average path length
apl = g.average_path_length()
print("Average path length: {:.3f}".format(apl))

# Get and save adjacency matrix (as sparse CSC matrix)
adj_matrix_sparse_fn = '../data/adjacency_matrix_sparce.npz'
adj_matrix_sparse = g.get_adjacency_sparse()
sp.sparse.save_npz(adj_matrix_sparse_fn, adj_matrix_sparse) # save CSC matrix
#sparse_matrix = sp.sparse.load_npz(adj_matrix_sparse_fn) # load CSC matrix

# Get graph diameter
diam = g.diameter()
print("Diameter of graph:", diam)

# Get degree centrality of graph
dc_fn = '../data/degree_centralities.csv'
dc = g.degree()
print("DEGREE CENTRALITY")
print("len: {:_}".format(len(dc)))
print("max: {:_}".format(max(dc)))
print("min: {}".format(min(dc)))
print("mean: {:.1f}".format(np.mean(dc)))
dc_df = pd.DataFrame(dc)
dc_df.to_csv(dc_fn, index=False, header=False)
plt.hist(dc, bins=50)
plt.show()
