==> notebooks/3.TEMPLATE Graph Between Timeframe.ipynb <==
# Functions: Analyse the evoltion of average positive and negative
import numpy as np
import pandas as pd
import dask.dataframe as dd
import igraph as ig
import matplotlib.pyplot as plt
import time
from fun.fun import *

pd.set_option('display.width', 500)

# Filenames
dataset_fn = "../dataset/TweetsCOV19.tsv"
plot_fn = "../images/plots/transitivity_evolution.png"

# -> IN : Read Edge List
edges_fn = "../data/edges.parquet"
print("reading edges ... ", end='')
start = time.time()
df = dd.read_parquet("../data/edges.parquet")
df[['source', 'target']] = df[['source', 'target']].astype('int32')
end = time.time()
print("read {:_} lines (took {:.1f}s)".format(len(df), (end-start)))
print(df.dtypes)
df.head()

# -> IN : Load tweets dataframe
tw = get_filtered_tweets_dataframe(dataset_fn)
print("Loaded {:_} tweets".format(len(tw)))
tw.head()

# -> IN : Read list of nodes
nodes = pd.read_csv('../data/node_ids.csv', index_col='twitter id')
print("Number of nodes: {:_}".format(len(nodes)))
nodes.head()

# Compute timestamp bins
steps = 100
ts_min, ts_max = min(tw['Timestamp']), max(tw['Timestamp'])
ts_inc = (ts_max - ts_min) / 100
ts_bins = [ ts_min + i*ts_inc for i in range(steps+1) ]

print("Min timestamp:", ts_min)
print("Max timestamp:", ts_max)
print("Timestamp increment:", ts_inc)

# Limit tweet ids to between timeframe and fetch index values
tw_sel = pd.DataFrame()
tw_sel['ids'] = np.array( tw[ (ts_bins[1] <= tw['Timestamp']) & (tw['Timestamp'] < ts_bins[2]) ].index )
tw_sel['ids'] = tw_sel['ids'].astype('int64')
tw_sel = tw_sel[tw_sel['ids'].isin(nodes.index)]
tw_sel['index'] = tw_sel['ids'].apply( lambda id: nodes.loc[id]['index'] )
tw_sel['index'] = tw_sel['index'].astype('int32')
#tw_sel.set_index('ids', inplace=True)
print(len(tw_sel))
node_index = { node: index for index, node in enumerate( sorted(tw_sel['index'].tolist()) ) }
print(node_index.values())

# fetch edges of nodes
edges = df[ (df['source'].isin(tw_sel['index'])) & (df['target'].isin(tw_sel['index'])) ].compute()
edges['source_i'] = edges['source'].apply( lambda x: node_index[x] )
edges['target_i'] = edges['target'].apply( lambda x: node_index[x] )
edges.head()
#edges.values.tolist()

df = edges.copy()
df.drop(['source_i', 'target_i'], axis=1, inplace=True)
adjls = df.groupby('source')['target'].apply(list, meta=('target', 'object')).compute()
adjls

indeces = tw_sel['index'].tolist()
for s, t in edges.values.tolist():
    #print(s, t)
    if s not in indeces or t not in indeces:
        print(s)
        break
    #print(t in tw_sel['index'])
    #print( not (s in tw_sel['index'] and t in tw_sel['index']) )
    #break
    #if not (s in tw_sel['index'] and t in tw_sel['index']):
    #    print(s, t)
    
print("done")

# iGraph
g = ig.Graph()
#vertices = tw_sel['index'].tolist()
g.add_vertices(len(node_index))
g.add_edges( edges[['source_i', 'target_i']].values.tolist() )
#g.add_edges(edges.values.tolist())
ig.summary(g)
gt = g.transitivity_undirected()
print(gt)

g = ig.Graph()
nodes = [2,3,4,6,7,8,9]
ss = [3,4,6,7,8]
ts = [2,3,9,8,9]
node_index = {node: index for index, node in enumerate(nodes)}
edges = pd.DataFrame(data={'source' : ss, 'target' : ts})
edges['source_i'] = edges['source'].apply( lambda x: node_index[x] )
edges['target_i'] = edges['target'].apply( lambda x: node_index[x] )
edges
g.add_vertices(len(nodes))
g.add_edges(edges[['source_i', 'target_i']].values.tolist())
print(g)
